Hard-code rotated secret that Slack used, for test
We want to ensure that a request Slack actually sent us continues to
work. For that, we need the secret they used to sign the request.

It's more convenient and secure to not have to set the current
production secret in e.g. CI, so we're hardcoding it here. It's rotated.
You can't use it anymore.