Add a price converter
- This will take an unparsed string and try it's best to extract the
currency out of the string. Right now the list of valid currencies is
hard coded - we may want to do this in a more generic way

- We will also need to change our domain now that price will contain
currency info. We also need to think about storing currency on the order
and whether we also need to store it on the order-item (the child)

[#168074457]